refactor(node-properties): delete PhysicalPedestrian base implementation
