+++ alchemist-physical-agents/src/main/kotlin/it/unibo/alchemist/model/implementations/properties/PhysicalPedestrian2D.kt
@@ -20,6 +20,8 @@ import it.unibo.alchemist.model.interfaces.environments.PhysicsEnvironment
 import it.unibo.alchemist.model.interfaces.geometry.euclidean2d.Euclidean2DShape
 import it.unibo.alchemist.model.interfaces.geometry.euclidean2d.Euclidean2DShapeFactory
 import it.unibo.alchemist.model.interfaces.geometry.euclidean2d.Euclidean2DTransformation
+import it.unibo.alchemist.model.interfaces.properties.CognitiveProperty
+import it.unibo.alchemist.nextDouble
 import org.apache.commons.math3.random.RandomGenerator
 
 /**
@@ -30,19 +32,40 @@ class PhysicalPedestrian2D<T>(
     /**
      * The environment in which the node is moving.
      */
-    override val environment: Physics2DEnvironment<T>,
-    node: Node<T>,
-) : PhysicalPedestrian<T, Euclidean2DPosition, Euclidean2DTransformation, Euclidean2DShapeFactory>(
-    randomGenerator,
-    environment,
-    node,
-),
-    PhysicalPedestrian2D<T> {
+    val environment: Physics2DEnvironment<T>,
+    override val node: Node<T>,
+) : PhysicalPedestrian2D<T> {
+
+    private val nodeShape by lazy { node.asProperty<T, AreaProperty<T>>().shape }
+
+    private val desiredSpaceTreshold: Double = randomGenerator.nextDouble(minimumSpaceTreshold, maximumSpaceThreshold)
+
+    override val comfortRay: Double get() {
+        val cognitiveModel = node.asPropertyOrNull<T, CognitiveProperty<T>>()?.cognitiveModel
+        return if (cognitiveModel?.wantsToEscape() == true) {
+            desiredSpaceTreshold / 3
+        } else {
+            desiredSpaceTreshold
+        }
+    }
 
     override val comfortArea: Euclidean2DShape get() = environment
         .shapeFactory
         .circle(node.asProperty<T, AreaProperty<T>>().shape.radius + comfortRay)
         .transformed { origin(environment.getPosition(node)) }
+
+    override fun repulsionForce(other: Node<T>): Euclidean2DPosition {
+        val myShape = nodeShape.transformed { origin(environment.getPosition(node)) }
+        val otherShape = environment.getShape(other)
+        return (myShape.centroid - otherShape.centroid).let {
+            val desiredDistance = myShape.radius + comfortRay + otherShape.radius
+            /*
+             * it is the vector leading from the center of other to the center of this node, it.magnitude is the
+             * actual distance between the two nodes.
+             */
+            it.normalized() * (desiredDistance - it.magnitude).coerceAtLeast(0.0) / it.magnitude
+        }
+    }
 
     override fun physicalForces(
         environment: PhysicsEnvironment<T, Euclidean2DPosition, Euclidean2DTransformation, Euclidean2DShapeFactory>,
@@ -58,4 +81,15 @@ class PhysicalPedestrian2D<T>(
         .toList()
 
     override fun cloneOnNewNode(node: Node<T>) = PhysicalPedestrian2D(randomGenerator, environment, node)
+
+    companion object {
+        /**
+         * Minimum value for normal state [comfortRay].
+         */
+        private const val minimumSpaceTreshold = 0.1
+        /**
+         * Maximum value for normal state [comfortRay].
+         */
+        private const val maximumSpaceThreshold = 1.0
+    }
 }
